chassis: Add library function to validate fabric is up
After upgrading the software on a portion of the chassis it is a good idea to
check if the upgrade succeeded before proceeding to upgrade other portions of
the chassis. This commit adds a check_fabric library function which looks to
see if BGP sessions are established on all of the fabric interfaces of the
line card or fabric card that was upgraded. This provides a minimal level of
comfort that the upgrade succeeded.